requirements/requirements.txt
asgiref==3.6.0
Django==3.2.18
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
django-environ==0.10.0
psycopg2==2.9.6
django-cleanup==7.0.0
sorl-thumbnail==12.9.0
pillow==9.5.0
celery==5.2.7
redis==4.5.4
django-webpush==0.3.5
channels==4.0.0
channels-redis==4.1.0
uvicorn[standard]==0.21.1
whitenoise==6.4.0